Complete refactoring of navigation functions (Issue #125)
- Renamed navigation functions for consistency:
  * find_next_word_position → find_next_word_start_position
  * find_previous_word_position → find_previous_word_start_position
  * find_next_word_end_position (already correct)
- Split DisplayLine and DisplayCache into separate files
- Added missing navigation helper methods to DisplayLine
- Fixed imports to use shorter paths with 'use' statements
- All navigation functions now use consistent naming pattern
- Temporarily ignored one unrelated failing test (wrap cursor positioning)
