+++ src/repl/models/mod.rs
@@ -9,6 +9,7 @@ pub mod buffer_char;
 pub mod buffer_model;
 pub mod display_cache;
 pub mod display_char;
+pub mod display_line;
 pub mod request_model;
 pub mod response_model;
 pub mod status_line;
@@ -16,8 +17,9 @@ pub mod status_line;
 // Re-export all models for easy access
 pub use buffer_char::{BufferChar, BufferLine, CharacterBuffer};
 pub use buffer_model::{BufferContent, BufferModel};
-pub use display_cache::{build_display_cache, DisplayCache, DisplayLine, DisplayPosition};
+pub use display_cache::{build_display_cache, DisplayCache, DisplayPosition};
 pub use display_char::DisplayChar;
+pub use display_line::DisplayLine;
 pub use request_model::{HttpHeaders, RequestModel};
 pub use response_model::ResponseModel;
 pub use status_line::{HttpStatus, StatusLine};
