+++ src/repl/view_models/pane_manager.rs
@@ -931,7 +931,7 @@ impl PaneManager {
         let current_display_pos = self.get_current_display_cursor();
 
         if let Some(new_pos) =
-            self.panes[self.current_pane].find_next_word_position(current_display_pos)
+            self.panes[self.current_pane].find_next_word_start_position(current_display_pos)
         {
             let events = self.set_current_display_cursor(new_pos);
             let mut all_events = events;
@@ -947,7 +947,7 @@ impl PaneManager {
         let current_display_pos = self.get_current_display_cursor();
 
         if let Some(new_pos) =
-            self.panes[self.current_pane].find_previous_word_position(current_display_pos)
+            self.panes[self.current_pane].find_previous_word_start_position(current_display_pos)
         {
             let events = self.set_current_display_cursor(new_pos);
             let mut all_events = events;
